hold

+++ apps/api/src/api/internal/webhooks.ts
@@ -1,10 +1,11 @@
+import { validate } from "@api/middleware/validate";
+import { Application } from "@application";
 import { IntegrationService } from "@domain/integration";
 import { Webhooks } from "@octokit/webhooks";
-import type { InstallationEvent } from "@octokit/webhooks-types";
+import type { InstallationEvent, PushEvent } from "@octokit/webhooks-types";
 import { Config } from "@services/config";
 import type { Member } from "@services/db";
 import { JWT } from "@services/jwt";
-import { validator } from "@services/validator";
 import { notOk, ok } from "@utils/network";
 import { Hono } from "hono";
 import { okAsync } from "neverthrow";
@@ -23,36 +24,58 @@ export const WebhooksRouter = new Hono()
       return notOk(c, { message: "Unauthorized" }, 401);
     }
 
-    const payload = await c.req.json<InstallationEvent>();
+    const event = c.req.header("X-GitHub-Event");
 
-    if (payload.action === "created") {
-      return IntegrationService.addGithubRepositoriesToIntegration(
-        payload.installation.id,
-        (payload.repositories || []).map((repository) => ({
-          id: repository.id,
-          name: repository.name,
-          owner: payload.installation.account.login,
-        }))
-      ).match(
-        (repositories) => ok(c, repositories),
-        (error) => notOk(c, { message: error.message }, 500)
-      );
+    if (event === "installation") {
+      const payload = await c.req.json<InstallationEvent>();
+
+      if (payload.action === "created") {
+        return IntegrationService.addGithubRepositoriesToIntegration(
+          payload.installation.id,
+          (payload.repositories || []).map((repository) => ({
+            id: repository.id,
+            name: repository.name,
+            owner: payload.installation.account.login,
+          }))
+        ).match(
+          (repositories) => ok(c, repositories),
+          (error) => notOk(c, { message: error.message }, 500)
+        );
+      }
+
+      if (payload.action === "deleted") {
+        return IntegrationService.deleteGithubIntegrationByInstallationId(
+          payload.installation.id
+        ).match(
+          (integration) => ok(c, integration),
+          (error) => notOk(c, { message: error.message }, 500)
+        );
+      }
     }
 
-    if (payload.action === "deleted") {
-      return IntegrationService.deleteGithubIntegrationByInstallationId(
-        payload.installation.id
-      ).match(
-        (integration) => ok(c, integration),
-        (error) => notOk(c, { message: error.message }, 500)
-      );
+    if (event === "push") {
+      const payload = await c.req.json<PushEvent>();
+      const headCommit = payload.head_commit;
+      const installation = payload.installation;
+
+      if (headCommit && installation) {
+        return Application.createPendingCommit(
+          headCommit.id,
+          installation.id,
+          payload.repository.owner.login,
+          payload.repository.name
+        ).match(
+          (integration) => ok(c, integration),
+          (error) => notOk(c, { message: error.message }, 500)
+        );
+      }
     }
 
     return ok(c, undefined);
   })
   .get(
     "/github/setup",
-    validator(
+    validate(
       "query",
       z.object({
         installation_id: z.number({ coerce: true }),
